requirements.txt
absl-py==2.3.1
alembic==1.13.3
annotated-types==0.7.0
anyio==4.10.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asttokens==3.0.0
astunparse==1.6.3
attrs==25.4.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.13.4
biopython==1.86
blinker==1.8.2
blis==1.3.0
build==1.3.0
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
cffi==2.0.0
chardet==3.0.4
charset-normalizer==3.3.2
chromadb==1.3.5
click==8.1.7
cloudpathlib==0.21.1
cloudscraper==1.2.71
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
confection==0.1.5
cosine-similarity==0.1.2
cryptography==46.0.3
cssselect==1.3.0
cymem==2.0.11
debugpy==1.8.11
decorator==5.1.1
deepface==0.0.95
distro==1.9.0
dnspython==2.8.0
durationpy==0.10
ecdsa==0.19.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
executing==2.1.0
fastapi==0.109.0
feedfinder2==0.0.4
feedparser==6.0.12
filelock==3.16.1
fire==0.7.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
flatbuffers==25.2.10
fsspec==2024.9.0
gast==0.6.0
gdown==5.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.65.0
googletrans==4.0.0rc1
greenlet==3.1.1
groq==0.36.0
grpcio==1.66.2
grpcio-status==1.66.2
gTTS==2.5.4
gunicorn==23.0.0
h11==0.16.0
h2==3.2.0
h5py==3.14.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==1.0.9
httplib2==0.22.0
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.25.1
humanfriendly==10.0
hyperframe==5.2.0
idna==2.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.30.0
itsdangerous==2.2.0
jedi==0.19.2
jieba3k==0.35.1
Jinja2==3.1.4
jiter==0.11.0
joblib==1.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.11.3
kubernetes==34.1.0
langcodes==3.5.0
langdetect==1.0.9
language_data==1.3.0
libclang==18.1.1
lxml==6.0.2
lxml_html_clean==0.4.3
lz4==4.4.4
Mako==1.3.5
marisa-trie==1.2.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
minio==7.2.3
ml_dtypes==0.5.3
mmh3==5.2.0
mpmath==1.3.0
mtcnn==1.0.0
murmurhash==1.0.13
mysql-connector-python==9.4.0
mysqlclient==2.2.7
namex==0.1.0
nest-asyncio==1.6.0
networkx==3.3
newspaper3k==0.2.8
nltk==3.9.2
numpy==2.1.2
oauthlib==3.3.1
onnxruntime==1.23.2
openai==1.109.1
opencv-python==4.12.0.88
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
opt_einsum==3.4.0
optree==0.17.0
orjson==3.11.4
overrides==7.7.0
packaging==24.1
pandas==2.3.2
parso==0.8.4
passlib==1.7.4
pillow==11.3.0
platformdirs==4.3.6
posthog==5.4.0
preshed==3.0.10
prompt_toolkit==3.0.48
proto-plus==1.24.0
protobuf==5.29.5
psutil==6.1.0
psycopg2-binary==2.9.9
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybase64==1.4.2
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
Pygments==2.18.0
PyJWT==2.9.0
pymongo==4.15.4
PyMuPDF==1.26.3
PyMySQL==1.1.2
pyparsing==3.1.4
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2025.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
redis==5.0.1
referencing==0.37.0
regex==2024.9.11
requests==2.32.3
requests-file==3.0.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retina-face==0.0.17
rfc3986==1.5.0
rich==14.1.0
rpds-py==0.29.0
rsa==4.9
safetensors==0.4.5
scikit-learn==1.7.1
scipy==1.16.1
sentence-transformers==5.1.0
setuptools==75.1.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
sniffio==1.3.1
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.14.3
SQLAlchemy==2.0.25
srsly==2.5.1
stack-data==0.6.3
starlette==0.35.1
sympy==1.13.3
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
tf_keras==2.20.1
thinc==8.3.6
threadpoolctl==3.6.0
tinysegmenter==0.3
tldextract==5.3.0
tokenizers==0.20.0
torch==2.4.1
tornado==6.4.2
tqdm==4.66.5
traitlets==5.14.3
transformers==4.45.1
typer==0.16.0
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.2.3
uv==0.5.7
uvicorn==0.27.0
wasabi==1.1.3
watchfiles==1.1.1
wcwidth==0.2.13
weasel==0.4.1
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.0.4
wheel==0.45.1
wrapt==1.17.3
zipp==3.23.0
